benchmarks/STRIDE-QA-Bench/pyproject.toml
[project]
name = "strideqa-bench"
version = "0.1.0"
description = "STRIDE-QA Bench"
requires-python = ">=3.10,<3.11"
dependencies = [
    "accelerate==1.6.0",
    "av==14.3.0",
    "bitsandbytes==0.45.5",
    "deepspeed==0.16.7",
    "detectron2", # This may fail on some systems. Try building it from source
    "dotenv==0.9.9",
    "einops==0.8.1",
    "huggingface-hub==0.30.2",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "Markdown==3.8.2",
    "markdown-it-py==3.0.0",
    "matplotlib-inline==0.1.7",
    "numpy==2.1.2",
    "omegaconf==2.3.0",
    "openai>=1.97.1",
    "opencv-python==4.11.0.86",
    "pandas==2.2.3",
    "peft==0.15.2",
    "pillow==11.0.0",
    "pycocotools==2.0.10",
    "python-dotenv==1.1.1",
    "PyYAML==6.0.2",
    "qwen-vl-utils==0.0.11",
    "rich==14.0.0",
    "safetensors==0.5.3",
    "seaborn==0.13.2",
    "sentencepiece>=0.2.0",
    "setuptools>=80.9.0",
    "tabulate==0.9.0",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "tensorboardX==2.6.2.2",
    "timm>=1.0.17",
    "tokenizers==0.21.1",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "transformers==4.51.3",
    "triton==3.2.0",
    "trl==0.17.0",
]

[project.optional-dependencies]
flash-attn = [
    "flash-attn>=2.7.4"
]

[dependency-groups]
dev = [
    "ruff",
    "ipykernel",
    "ipywidgets",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where   = ["src"]
include = ["strideqa_bench*"]

[tool.uv]
no-build-isolation-package = ["flash-attn", "detectron2"]
index-strategy = "unsafe-best-match"
environments = [
  "sys_platform == 'linux' and platform_machine == 'x86_64'"
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]
torchvision = [
    { index = "pytorch-cu124", marker = "sys_platform == 'linux'" },
]
detectron2 = [
    { git = "https://github.com/facebookresearch/detectron2.git", rev = "18f69583391e5040043ca4f4bebd2c60f0ebfde0" },
]
flash-attn = [
  { url = "https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.1.post4/flash_attn-2.7.1.post4+cu12torch2.6cxx11abiFALSE-cp310-cp310-linux_x86_64.whl" },
]

[[tool.uv.index]]
name     = "pytorch-cu124"
url      = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.ruff]
src = ["strideqa_bench"]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
target-version = "py312"
line-length = 100

[tool.ruff.lint.isort]
known-first-party = ["strideqa_bench"]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
